Move screen components from activity to screen package
